Fix categories
