+++ styles/blog.scss
@@ -18,6 +18,7 @@
   flex-grow: 1;
   margin-top: 0.625rem;
   word-wrap: anywhere;
+  margin-bottom: 1.5rem;
 
   @media screen and (max-width: $tablet) {
     padding: 0 5%;
